+++ crates/zng-task/src/ipc.rs
@@ -97,6 +97,8 @@ use zng_unit::TimeUnits as _;
 #[doc(no_inline)]
 pub use ipc_channel::ipc::{IpcBytesReceiver, IpcBytesSender, IpcReceiver, IpcSender, bytes_channel};
 
+use crate::TaskPanicError;
+
 /// Represents a type that can be an input and output of IPC workers.
 ///
 /// # Trait Alias
@@ -290,7 +292,7 @@ impl<I: IpcValue, O: IpcValue> Worker<I, O> {
             match crate::with_deadline(crate::wait(move || receiver.join()), 1.secs()).await {
                 Ok(r) => {
                     if let Err(p) = r {
-                        tracing::error!("worker receiver thread exited panicked, {}", crate::crate_util::panic_str(&p));
+                        tracing::error!("worker receiver thread exited panicked, {}", TaskPanicError::new(p).panic_str().unwrap_or(""));
                     }
                 }
                 Err(_) => {
@@ -353,7 +355,7 @@ impl<I: IpcValue, O: IpcValue> Worker<I, O> {
             let (t, p) = self.running.take().unwrap();
 
             if let Err(e) = t.join() {
-                tracing::error!("panic in worker receiver thread, {}", crate::crate_util::panic_str(&e));
+                tracing::error!("panic in worker receiver thread, {}", TaskPanicError::new(e).panic_str().unwrap_or(""));
             }
 
             if let Err(e) = p.kill() {