+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     - **Breaking** Remove `zng::color::RenderMixBlendMode`.
     - **Breaking** `zng::color::MixBlendMode` is now non-exhaustive and has a new variant `PlusLighter`.
 
+* **Breaking** Task functions that capture panic now return `zng::task::TaskPanicError`.
 * Fix gradient stops with midway adjustment.
 * Impl of `Add` and `Sub` for `layout::Vector` is now generic over any type that converts to vector.
 * Add `Var::chase_begin` to begin a deferred chase animation.